Коммит задания из 5 подмодуля, модуляция
